docker/cpu/pyproject.toml
[project]
name = "minutes-maker"
version = "0.1.0"
description = "A tool to make minutes of meeting automatically."
authors = [{ name = "discus0434", email = "discus0434@gmail.com" }]
dependencies = [
    "openai~=0.27.8",
    "tiktoken~=0.4.0",
    "faster-whisper~=0.6.0",
    "setuptools>=68.0.0",
    "python-dotenv>=1.0.0",
    "fastapi[uvicorn]~=0.99.0",
    "uvicorn~=0.22.0",
    "python-multipart~=0.0.6",
    "pydub>=0.25.1", 
    "assemblyai~=0.41.3"
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.3.0",
    "isort~=5.12.0",
    "ipykernel~=6.23.3",
    "mypy~=1.4.1",
    "flake8~=6.0.0",
]
[tool.hatch.metadata]
allow-direct-references = true
